Consolidate multiversion code into a single file.
